requirements.txt
flask
pytest
selenium
allure-pytest